More refactor into db.go
